+++ spacer/storage.py
@@ -22,6 +22,7 @@ from sklearn.calibration import CalibratedClassifierCV
 from sklearn.linear_model import SGDClassifier
 
 from spacer import config
+from spacer.aws import get_s3_resource
 from spacer.exceptions import URLDownloadError
 
 
@@ -161,27 +162,42 @@ class S3Storage(RemoteStorage):
         self.transfer_config = TransferConfig(use_threads=False)
 
     def store(self, key: str, stream: BytesIO):
-        s3 = config.get_s3_resource()
+        s3 = get_s3_resource()
         s3.Bucket(self.bucket_name).put_object(Body=stream, Key=key)
 
     def _load_remote(self, key: str):
-        s3 = config.get_s3_resource()
+        s3 = get_s3_resource()
         stream = BytesIO()
         s3.Object(self.bucket_name, key).download_fileobj(
             stream, Config=self.transfer_config)
         return stream
 
     def delete(self, key: str) -> None:
-        s3 = config.get_s3_resource()
+        s3 = get_s3_resource()
         s3.Object(self.bucket_name, key).delete()
 
     def exists(self, key: str):
-        s3 = config.get_s3_resource()
+        """
+        The error-handling logic is from django-storages'
+        S3Storage.exists().
+        """
+        s3 = get_s3_resource()
         try:
+            # This calls head_object() to retrieve object metadata
+            # without retrieving the object itself.
+            # https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3/object/load.html
             s3.Object(self.bucket_name, key).load()
             return True
-        except botocore.exceptions.ClientError:
-            return False
+        except botocore.exceptions.ClientError as err:
+            if err.response["ResponseMetadata"]["HTTPStatusCode"] == 404:
+                # The object doesn't exist AND we have the ListBucket
+                # permission (which we expect to).
+                # If we can't ListBucket, it'll be 403 instead.
+                # https://repost.aws/articles/ARe3OTZ3SCTWWqGtiJ6aHn8Q/why-does-s-3-return-403-instead-of-404-when-the-object-doesnt-exist
+                return False
+
+            # Some other error was encountered. Re-raise it.
+            raise
 
 
 class FileSystemStorage(Storage):