+++ spacer/tests/test_data_classes.py
@@ -4,6 +4,7 @@ import unittest
 import random
 
 from spacer import config
+from spacer.aws import get_s3_resource
 from spacer.data_classes import \
     PointFeatures, \
     ImageFeatures, \
@@ -75,7 +76,7 @@ class TestImageFeaturesNumpyStore(unittest.TestCase):
 
         self._test_numpy_store(s3_loc)
 
-        s3 = config.get_s3_resource()
+        s3 = get_s3_resource()
         s3.Object(config.TEST_BUCKET, s3_loc.key).delete()
 
     def test_fs(self):