+++ CHANGELOG.md
@@ -10,6 +10,13 @@
 
 - Fixed an `AttributeError` in `TorchExtractor.load_weights()` that would happen on numpy 1.26.1 through 1.26.5. These are the versions where `numpy._core` exists but doesn't have attributes available; on these versions we now access `numpy.core` like on other 1.x versions.
 
+- AWS related changes:
+
+  - Added `AWS_PROFILE_NAME` and `AWS_SESSION_TOKEN` config variables to support more ways of accessing AWS.
+  - Moved `get_s3_resource()` from `config.py` to a new file, `aws.py`.
+  - Added an `aws_check()` function in `aws.py`, to help debug AWS configuration.
+  - `S3Storage.exists()` now only returns False if there is a ClientError with status code 404. In other ClientError cases, including 403, this method will now re-raise the error. Due to this, it's now recommended to ensure you have the ListBucket permission for any S3 bucket you work with; otherwise, S3 will return 403 instead of 404 for a missing file (see [this article](https://repost.aws/articles/ARe3OTZ3SCTWWqGtiJ6aHn8Q/why-does-s-3-return-403-instead-of-404-when-the-object-doesnt-exist)).
+
 ## 0.12.0
 
 - Updates to pip-install dependencies: